second 50

+++ LeetCodeCSharp/Ex014-LongestCommonPrefix.cs
@@ -23,8 +23,8 @@ namespace LeetCodeCSharp
         {
             // brute force O(mn)
             if (strs.Count == 0) return "";
-            for (int i = 0; i < (int)strs[0].Length; i++)
-                for (int j = 1; j < (int)strs.Count; j++)
+            for (int i = 0; i < strs[0].Length; i++)
+                for (int j = 1; j < strs.Count; j++)
                     if (strs[j][i] != strs[0][i]) return strs[0].Substring(0, i);
             return strs[0];
         }
