+++ LeetCodeCSharp/Ex038-CombinationSum-Test.cs
@@ -13,7 +13,7 @@ namespace LeetCodeCSharp
             List<int> t = new List<int>();
             t.Add(2); t.Add(3); t.Add(6); t.Add(7);
             List<List<int>> res = Ex038.CombinationSum(t, 7);
-            Assert.AreEqual(2, (int)res.Count);
+            Assert.AreEqual(2, res.Count);
         }
 
         [TestMethod]
@@ -22,7 +22,7 @@ namespace LeetCodeCSharp
             List<int> t = new List<int>();
             t.Add(2); t.Add(3); t.Add(6); t.Add(7);
             List<List<int>> res = Ex038.CombinationSum(t, 6);
-            Assert.AreEqual(3, (int)res.Count);
+            Assert.AreEqual(3, res.Count);
         }
 
         [TestMethod]
@@ -31,7 +31,7 @@ namespace LeetCodeCSharp
             List<int> t = new List<int>();
             t.Add(2); t.Add(3); t.Add(6); t.Add(7);
             List<List<int>> res = Ex038.CombinationSum(t, 1);
-            Assert.AreEqual(0, (int)res.Count);
+            Assert.AreEqual(0, res.Count);
         }
     };
 }
