+++ LeetCodeCSharp/Ex039-CombinationSumII-Test.cs
@@ -13,7 +13,7 @@ namespace LeetCodeCSharp
             List<int> t = new List<int>();
             t.Add(2); t.Add(3); t.Add(6); t.Add(7);
             List<List<int>> res = Ex039.CombinationSum2(t, 7);
-            Assert.AreEqual(1, (int)res.Count);
+            Assert.AreEqual(1, res.Count);
         }
 
         [TestMethod]
@@ -22,7 +22,7 @@ namespace LeetCodeCSharp
             List<int> t = new List<int>();
             t.Add(2); t.Add(3); t.Add(6); t.Add(7);
             List<List<int>> res = Ex039.CombinationSum2(t, 5);
-            Assert.AreEqual(1, (int)res.Count);
+            Assert.AreEqual(1, res.Count);
         }
 
         [TestMethod]
@@ -32,7 +32,7 @@ namespace LeetCodeCSharp
             t.Add(2); t.Add(3); t.Add(6); t.Add(7);
             t.Add(1); t.Add(1);
             List<List<int>> res = Ex039.CombinationSum2(t, 6);
-            Assert.AreEqual(2, (int)res.Count);
+            Assert.AreEqual(2, res.Count);
         }
     };
 }
