+++ LeetCodeCSharp/Ex050-N-Queens-Test.cs
@@ -11,28 +11,28 @@ namespace LeetCodeCSharp
         public void LeetCode_Ex050_Test_SolveNQueens1()
         {
             List<List<string>> res = Ex050.SolveNQueens(4);
-            Assert.AreEqual(2, (int)res.Count);
+            Assert.AreEqual(2, res.Count);
         }
 
         [TestMethod]
         public void LeetCode_Ex050_Test_SolveNQueens2()
         {
             List<List<string>> res = Ex050.SolveNQueens(2);
-            Assert.AreEqual(0, (int)res.Count);
+            Assert.AreEqual(0, res.Count);
         }
 
         [TestMethod]
         public void LeetCode_Ex050_Test_SolveNQueens3()
         {
             List<List<string>> res = Ex050.SolveNQueens(3);
-            Assert.AreEqual(0, (int)res.Count);
+            Assert.AreEqual(0, res.Count);
         }
 
         [TestMethod]
         public void LeetCode_Ex050_Test_SolveNQueens4()
         {
             List<List<string>> res = Ex050.SolveNQueens(5);
-            Assert.AreEqual(10, (int)res.Count);
+            Assert.AreEqual(10, res.Count);
         }
     };
 }
